Move implementation of Peer into header file to break dependency on libNetwork.
